+++ CotEditor/Sources/NSLayoutManager+InvisibleDrawing.swift
@@ -79,6 +79,11 @@ extension InvisibleDrawing {
             
             let glyphIndex = self.glyphIndexForCharacter(at: charIndex)
             
+            var lineFragmentRange: NSRange = .notFound
+            let lineOrigin = self.lineFragmentRect(forGlyphAt: glyphIndex, effectiveRange: &lineFragmentRange, withoutAdditionalLayout: true).origin
+            let glyphLocation = self.location(forGlyphAt: glyphIndex)
+            let symbolOrigin = lineOrigin.offset(by: origin).offsetBy(dx: glyphLocation.x, dy: baselineOffset - glyphHeight)
+            
             let path: NSBezierPath
             if let cache = pathCache[codeUnit] {
                 path = cache
@@ -95,7 +100,11 @@ extension InvisibleDrawing {
                         // for non-zeroAdvancement controls, such as VERTICAL TABULATION
                         glyphWidth = self.boundingBoxForControlGlyph(for: self.textFont).width
                     default:
-                        glyphWidth = self.enclosingRectForGlyph(at: glyphIndex, in: textContainer).width
+                        // -> Avoid invoking `.enclosingRectForGlyph(at:in:)` as much as possible
+                        //    that takes long time with long unwrapped lines.
+                        glyphWidth = lineFragmentRange.contains(glyphIndex + 1)
+                            ? self.location(forGlyphAt: glyphIndex + 1).x - glyphLocation.x
+                            : self.enclosingRectForGlyph(at: glyphIndex, in: textContainer).width
                 }
                 
                 let size = CGSize(width: glyphWidth, height: glyphHeight)
@@ -106,10 +115,6 @@ extension InvisibleDrawing {
                     pathCache[codeUnit] = path
                 }
             }
-            
-            let lineOrigin = self.lineFragmentRect(forGlyphAt: glyphIndex, effectiveRange: nil, withoutAdditionalLayout: true).origin
-            let glyphLocation = self.location(forGlyphAt: glyphIndex)
-            let symbolOrigin = lineOrigin.offset(by: origin).offsetBy(dx: glyphLocation.x, dy: baselineOffset - glyphHeight)
             
             path.transform(using: .init(translationByX: symbolOrigin.x, byY: symbolOrigin.y))
             path.fill()